+++ internal/monitor/command_test.go
